+++ CHANGELOG.md
@@ -2,6 +2,40 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.141](https://github.com/wrappid/styles/compare/v0.0.140...v0.0.141) (2024-03-23)
+
+
+### Features
+
+* **global:** :sparkles: added different bg-grid styles ([7f3d1d6](https://github.com/wrappid/styles/commit/7f3d1d66fdbc87fa4165c00f480643e586d27d0d)), closes [#57](https://github.com/wrappid/styles/issues/57)
+
+### [0.0.140](https://github.com/wrappid/styles/compare/v0.0.139...v0.0.140) (2024-03-22)
+
+
+### Features
+
+* **global:** :sparkles: initial changes to StylesProvider.tsx ([4960ca7](https://github.com/wrappid/styles/commit/4960ca74bd02a2ff653927cf3b756fafee532687)), closes [#57](https://github.com/wrappid/styles/issues/57)
+
+### [0.0.139](https://github.com/wrappid/styles/compare/v0.0.138...v0.0.139) (2024-03-21)
+
+
+### Bug Fixes
+
+* **core:** :bug: bug fixed for bg dotted background styles and a text styles css ([caec533](https://github.com/wrappid/styles/commit/caec53349348798a93ec050b912299398d5a6d9f)), closes [#72](https://github.com/wrappid/styles/issues/72)
+
+### [0.0.138](https://github.com/wrappid/styles/compare/v0.0.137...v0.0.138) (2024-03-20)
+
+
+### Features
+
+* **core:** :sparkles: created styles for transform rotate and a background ([35125c7](https://github.com/wrappid/styles/commit/35125c71bb873c88314d65222755735339b7a5b0)), closes [#72](https://github.com/wrappid/styles/issues/72)
+
+### 0.0.137 (2024-03-20)
+
+### 0.0.136 (2024-03-20)
+
+### 0.0.135 (2024-03-19)
+
 ### 0.0.134 (2024-03-18)
 
 ### 0.0.133 (2024-03-16)